ìì회우ë원가입폼추가

+++ src/main/java/com/example/Ecommerce/domain/siteuser/Repository/SiteuserRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SiteuserRepository extends JpaRepository<Siteuser, Integer> {
+public interface SiteuserRepository extends JpaRepository<Siteuser, Long> {
+
 }
